import_hostings_from_calc/pyproject.toml
[project]
name = "import-hostings-from-calc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ezodf>=0.3.2",
    "flask>=3.1.2",
    "flask-sqlalchemy>=3.1.1",
    "lxml>=6.0.2",
    "odfpy>=1.4.1",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "psycopg2-binary>=2.9.11",
    "pyexcel-ods>=0.6.0",
    "requests>=2.32.5",
]
